first swag at listing out clients
