+++ app/components/clients/list_item_component.html.erb
@@ -1,27 +1,21 @@
 <li class="relative flex justify-between gap-x-6 py-5">
   <div class="flex min-w-0 gap-x-4">
-    <%= helpers.user_gravatar(user: @user, css_classes: "h-12 w-12 flex-none rounded-full bg-gray-50") %>
     <div class="min-w-0 flex-auto">
       <p class="text-sm font-semibold leading-6 text-gray-900">
-        <%= link_to settings_user_path(@user) do %>
+        <%= link_to client_path(@client) do %>
           <span class="absolute inset-x-0 -top-px bottom-0"></span>
-          <%= user_name %>
+          <%= client_name %>
         <% end %>
       </p>
       <p class="mt-1 flex text-xs leading-5 text-gray-500">
-        <a href="mailto:<%= user_email %>" class="relative truncate hover:underline">
-          <%= user_email %>
-        </a>
+        <%= client_description %>
       </p>
     </div>
   </div>
   <div class="flex shrink-0 items-center gap-x-4">
     <div class="hidden sm:flex sm:flex-col sm:items-end">
-      <% if user_job_title %>
-        <p class="text-sm leading-6 text-gray-900"><%= user_job_title %></p>
-      <% end %>
       <div class="mt-1 flex items-center gap-x-1.5">
-        <%= render(Shared::StatusComponent.new(status: user_status)) %>
+        <%= render(Shared::StatusComponent.new(status: client_status)) %>
       </div>
     </div>
     <svg class="h-5 w-5 flex-none text-gray-400" viewBox="0 0 20 20" fill="currentColor" aria-hidden="true">
