--- app/components/settings/users/status_component.html.erb
@@ -1,3 +1,0 @@
-<span class="inline-flex items-center rounded-full px-2 py-1 text-xs font-medium ring-1 ring-inset <%= user_status_color %>">
-  Status: <%= user_status %>
-</span>
