+++ app/components/shared/status_component.html.erb
@@ -1,0 +1,3 @@
+<span class="inline-flex items-center rounded-full px-2 py-1 text-xs font-medium ring-1 ring-inset <%= target_status_color %>">
+  Status: <%= target_status.titleize %>
+</span>
