+++ app/views/clients/index.html.erb
@@ -12,3 +12,17 @@
     <%= render @clients %>
   </div>
 </div>
+
+<%= render(Settings::TabsComponent.new) do |component| %>
+  <% component.with_action_buttons do %>
+    <%= link_to "New Client", new_client_path, class: "ml-3 inline-flex items-center rounded-md bg-indigo-600 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600" %>
+  <% end %>
+<% end %>
+
+<div class="bg-white p-4 mt-4">
+  <ul role="list" class="divide-y divide-gray-100">
+    <% @clients.each do |client| %>
+      <%= render(Clients::ListItemComponent.new(client:, current_company:)) %>
+    <% end %>
+  </ul>
+</div>
